save/submit clickEventListener function in the EventsForm.js file also move the renderEvents function into the EventsForm.js file.
