+++ src/scripts/Events/EventsDataProvider.js
@@ -2,7 +2,6 @@
 
 const contentEventTarget = document.querySelector(".events");
 const eventHub = document.querySelector(".dashboard");
-// const listOfEventsURL = "";
 
 let eventsArray = [];
 
@@ -14,21 +13,21 @@ export const getEvents = () => {
     return fetch(`http://localhost:8088/events?_expand=user`)
         .then(response => response.json())
         .then(parsedEvents => {
-            events = 
-            parsedEvents
+            events =
+                parsedEvents
         })
-    }
+};
 
-    export const saveEvents = (eventObj) => {
-        return fetch(`http://localhost:8088/events?_expand=user`, {
-            method: "POST",
-            headers: {
-                "content-type":
+export const saveEvents = (eventObj) => {
+    return fetch(`http://localhost:8088/events?_expand=user`, {
+        method: "POST",
+        headers: {
+            "content-type":
                 "application/json"
-            },
-            body: JSON.stringify
+        },
+        body: JSON.stringify
             (eventObj)
-        })
+    })
         .then(getEvents)
         .then(dispatchEvent)
-    };
+};
